Fixed Bug related to Cell reused
Cells are reused for performance reasons but if you check mark a cell the check mark will also show up in another reused cell.  Added an array of checked values and then updated the array with true if they are checked.  Also added an If statement to make sure the accesorytype matches the array